Split off big chunk of websocket module into websocket_common module. And load the websocket_common module by default (which is just an API).
