+++ doc/conf/modules.default.conf
@@ -257,6 +257,7 @@ loadmodule "json-log-tag"; /* unrealircd.org/json-log tag and CAP for ircops */
 loadmodule "targetfloodprot"; /* set::anti-flood::target-flood protection */
 loadmodule "watch-backend"; /* used by watch and other modules */
 loadmodule "geoip_base"; /* needed for ALL geoip functions */
+loadmodule "websocket_common"; /* helper functions for websocket (internal) */
 
 loadmodule "geoip_classic";
 @if module-loaded("geoip_classic")
