+++ include/modules.h
@@ -2496,6 +2496,9 @@ enum EfunctionType {
 	EFUNC_RPC_RESPONSE,
 	EFUNC_RPC_ERROR,
 	EFUNC_RPC_ERROR_FMT,
+	EFUNC_WEBSOCKET_HANDLE_WEBSOCKET,
+	EFUNC_WEBSOCKET_CREATE_PACKET,
+	EFUNC_WEBSOCKET_CREATE_PACKET_SIMPLE,
 };
 
 /* Module flags */
