+++ src/api-efunctions.c
@@ -142,6 +142,9 @@ int (*webserver_handle_body)(Client *client, WebRequest *web, const char *readbu
 void (*rpc_response)(Client *client, json_t *request, json_t *result);
 void (*rpc_error)(Client *client, json_t *request, int error_code, const char *error_message);
 void (*rpc_error_fmt)(Client *client, json_t *request, int error_code, const char *fmt, ...);
+int (*websocket_handle_websocket)(Client *client, WebRequest *web, const char *readbuf2, int length2, int callback(Client *client, char *buf, int len));
+int (*websocket_create_packet)(int opcode, char **buf, int *len);
+int (*websocket_create_packet_simple)(int opcode, const char **buf, int *len);
 
 Efunction *EfunctionAddMain(Module *module, EfunctionType eftype, int (*func)(), void (*vfunc)(), void *(*pvfunc)(), char *(*stringfunc)(), const char *(*conststringfunc)())
 {
@@ -419,4 +422,7 @@ void efunctions_init(void)
 	efunc_init_function(EFUNC_RPC_RESPONSE, rpc_response, rpc_response_default_handler);
 	efunc_init_function(EFUNC_RPC_ERROR, rpc_error, rpc_error_default_handler);
 	efunc_init_function(EFUNC_RPC_ERROR_FMT, rpc_error_fmt, rpc_error_fmt_default_handler);
+	efunc_init_function(EFUNC_WEBSOCKET_HANDLE_WEBSOCKET, websocket_handle_websocket, websocket_handle_websocket_default_handler);
+	efunc_init_function(EFUNC_WEBSOCKET_CREATE_PACKET, websocket_create_packet, websocket_create_packet_default_handler);
+	efunc_init_function(EFUNC_WEBSOCKET_CREATE_PACKET_SIMPLE, websocket_create_packet_simple, websocket_create_packet_simple_default_handler);
 }
